Added fix for Silverlight/WP7 to use expression compilation instead of reflection to fetch object fields to get around their (obviously flawed) security model.
