+++ tests/CreateTableTest.cs
@@ -1,11 +1,5 @@
-
 using System;
-using System.IO;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using SQLite;
-
 using NUnit.Framework;
 
 namespace SQLite.Tests
@@ -13,51 +7,6 @@ namespace SQLite.Tests
 	[TestFixture]
 	public class CreateTableTest
 	{
-		public class Product
-		{
-			[AutoIncrement, PrimaryKey]
-			public int Id { get; set; }
-			public string Name { get; set; }
-			public decimal Price { get; set; }
-		}
-		public class Order
-		{
-			[AutoIncrement, PrimaryKey]
-			public int Id { get; set; }
-			public DateTime PlacedTime { get; set; }
-		}
-		public class OrderHistory {
-			[AutoIncrement, PrimaryKey]
-			public int Id { get; set; }
-			public int OrderId { get; set; }
-			public DateTime Time { get; set; }
-			public string Comment { get; set; }
-		}
-		public class OrderLine
-		{
-			[AutoIncrement, PrimaryKey]
-			public int Id { get; set; }
-            [Indexed("IX_OrderProduct", 1)]
-			public int OrderId { get; set; }
-            [Indexed("IX_OrderProduct", 2)]
-            public int ProductId { get; set; }
-			public int Quantity { get; set; }
-			public decimal UnitPrice { get; set; }
-			public OrderLineStatus Status { get; set; }
-		}
-		public enum OrderLineStatus {
-			Placed = 1,
-			Shipped = 100
-		}
-
-		public class TestDb : SQLiteConnection
-		{
-			public TestDb () : base(Path.GetTempFileName ())
-			{
-				Trace = true;
-			}
-		}
-
 		[Test]
 		public void CreateThem ()
 		{
@@ -75,7 +24,7 @@ namespace SQLite.Tests
 				Status = OrderLineStatus.Shipped
 			};
 			db.Insert(l);
-			var lo = db.Table<OrderLine>().Where(x => x.Status == CreateTableTest.OrderLineStatus.Shipped).FirstOrDefault();
+			var lo = db.Table<OrderLine>().Where(x => x.Status == OrderLineStatus.Shipped).FirstOrDefault();
 			Assert.AreEqual(lo.Id, l.Id);
 		}
 	}
